Refactor wordfreq/storage/database.py into smaller, focused modules
Split the large 1700+ line database.py file into smaller, purpose-based modules:

**New Structure:**
- `models/guid_prefixes.py` - GUID prefix mapping table
- `utils/guid.py` - GUID generation utilities
- `utils/enums.py` - POS enum helpers
- `utils/session.py` - Database session management
- `utils/initialization.py` - Corpus initialization
- `crud/word_token.py` - WordToken operations
- `crud/lemma.py` - Lemma operations
- `crud/derivative_form.py` - DerivativeForm operations
- `crud/word_frequency.py` - WordFrequency operations
- `crud/example_sentence.py` - ExampleSentence operations
- `crud/grammar_fact.py` - GrammarFact operations
- `queries/pos.py` - POS-based queries
- `queries/translation.py` - Translation queries
- `queries/stats.py` - Statistics queries
- `queries/noun_forms.py` - Noun form queries
- `legacy.py` - Backward compatibility functions

**Benefits:**
- Reduced database.py from 1709 lines to 251 lines (exports only)
- Improved maintainability with focused, single-purpose modules
- Full backward compatibility - all existing imports still work
- Better organization by separating CRUD, queries, and utilities

All callers continue to work without changes as database.py re-exports
everything for backward compatibility.
