+++ src/wordfreq/storage/crud/__init__.py
@@ -1,0 +1,1 @@
+"""CRUD operations for storage models."""
