+++ src/wordfreq/storage/queries/__init__.py
@@ -1,0 +1,1 @@
+"""Query functions for storage models."""
