+++ src/wordfreq/storage/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Storage utility functions."""
